requirements/base.txt
amqp
asgiref
attrs
billiard
celery
certifi
# Removed boto3 and botocore
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
Django
django-cors-headers
django-filter
# Removed django-storages (AWS-related)
django-tenants
djangorestframework
djangorestframework-simplejwt
drf-spectacular
feedparser
gunicorn
icalendar
idna
ijson
inflection
jmespath
jsonschema
jsonschema-specifications
kombu
mailersend
packaging
pika
pillow
prompt_toolkit
psycopg2-binary
pycurl
PyJWT
pymemcache
python-dateutil
python-decouple
pytz
PyYAML
referencing
requests
rpds-py
setuptools
sgmllib3k
six
sqlparse
supervisor
tenant-schemas-celery
tzdata
uritemplate
urllib3
vine
wcwidth
djangorestframework-simplejwt
django-celery-beat
flower

celery
redis
django-celery-beat
flower
psycopg2-binary
django-redis
python-dotenv
django-debug-toolbar